refactor: remove namespaced exports, fix package jsons and strip node protocol
