+++ package.json
@@ -8,6 +8,7 @@
   },
   "scripts": {
     "prepare": "husky",
+    "turbo:clean-cache": "rimraf ./**/.turbo/",
     "turbo": "turbo",
     "assist": "biome check --formatter-enabled=false --linter-enabled=false --assist-enabled=true --no-errors-on-unmatched",
     "assist:apply": "biome check --write --formatter-enabled=false --linter-enabled=false --assist-enabled=true --no-errors-on-unmatched",
@@ -30,6 +31,7 @@
     "@vitest/coverage-v8": "^3.2.4",
     "husky": "^9.1.7",
     "prettier": "^3.6.2",
+    "rimraf": "^6.0.1",
     "tsdown": "^0.12.9",
     "turbo": "^2.5.4",
     "type-fest": "^4.41.0",
