+++ packages/aio-commerce-lib-core/docs/api-reference/README.md
@@ -5,9 +5,3 @@
 | Module                   | Description                                                     |
 | ------------------------ | --------------------------------------------------------------- |
 | [error](error/README.md) | This module exports the error classes for the AIO Commerce SDK. |
-
-## References
-
-### Error
-
-Renames and re-exports [error](error/README.md)
