+++ packages/aio-commerce-lib-core/docs/api-reference/error/classes/CommerceSdkErrorBase.md
@@ -1,6 +1,6 @@
 # ``abstract` CommerceSdkErrorBase`
 
-Defined in: [lib/error/base-error.ts:53](https://github.com/adobe/aio-commerce-sdk/blob/b828858b2e024cee9599e664761b0da5b22f0cd1/packages/aio-commerce-lib-core/source/lib/error/base-error.ts#L53)
+Defined in: base-error.ts:53
 
 Base class for all the errors in the AIO Commerce SDK.
 
@@ -38,7 +38,7 @@ console.log(err.toJSON());
 new CommerceSdkErrorBase(message: string, options?: CommerceSdkErrorBaseOptions): CommerceSdkErrorBase;
 ```
 
-Defined in: [lib/error/base-error.ts:63](https://github.com/adobe/aio-commerce-sdk/blob/b828858b2e024cee9599e664761b0da5b22f0cd1/packages/aio-commerce-lib-core/source/lib/error/base-error.ts#L63)
+Defined in: base-error.ts:63
 
 Constructs a new CommerceSdkErrorBase instance. This is an abstract class so you
 should not instantiate it directly. Only invoke this constructor from a subclass.
@@ -68,7 +68,7 @@ Error.constructor;
 readonly optional traceId: string;
 ```
 
-Defined in: [lib/error/base-error.ts:54](https://github.com/adobe/aio-commerce-sdk/blob/b828858b2e024cee9599e664761b0da5b22f0cd1/packages/aio-commerce-lib-core/source/lib/error/base-error.ts#L54)
+Defined in: base-error.ts:54
 
 ## Accessors
 
@@ -80,7 +80,7 @@ Defined in: [lib/error/base-error.ts:54](https://github.com/adobe/aio-commerce-s
 get fullStack(): string;
 ```
 
-Defined in: [lib/error/base-error.ts:96](https://github.com/adobe/aio-commerce-sdk/blob/b828858b2e024cee9599e664761b0da5b22f0cd1/packages/aio-commerce-lib-core/source/lib/error/base-error.ts#L96)
+Defined in: base-error.ts:96
 
 Returns the full stack trace of the error and its causes.
 
@@ -98,7 +98,7 @@ Returns the full stack trace of the error and its causes.
 get rootCause(): unknown;
 ```
 
-Defined in: [lib/error/base-error.ts:109](https://github.com/adobe/aio-commerce-sdk/blob/b828858b2e024cee9599e664761b0da5b22f0cd1/packages/aio-commerce-lib-core/source/lib/error/base-error.ts#L109)
+Defined in: base-error.ts:109
 
 Returns the root cause of the error.
 
@@ -120,7 +120,7 @@ toJSON(): {
 };
 ```
 
-Defined in: [lib/error/base-error.ts:124](https://github.com/adobe/aio-commerce-sdk/blob/b828858b2e024cee9599e664761b0da5b22f0cd1/packages/aio-commerce-lib-core/source/lib/error/base-error.ts#L124)
+Defined in: base-error.ts:124
 
 Converts the error to a JSON-like representation.
 
@@ -174,7 +174,7 @@ traceId: undefined | string;
 toString(inspect: boolean): string;
 ```
 
-Defined in: [lib/error/base-error.ts:139](https://github.com/adobe/aio-commerce-sdk/blob/b828858b2e024cee9599e664761b0da5b22f0cd1/packages/aio-commerce-lib-core/source/lib/error/base-error.ts#L139)
+Defined in: base-error.ts:139
 
 Returns a pretty string representation of the error.
 
@@ -196,7 +196,7 @@ Returns a pretty string representation of the error.
 static isSdkError(error: unknown): error is CommerceSdkErrorBase;
 ```
 
-Defined in: [lib/error/base-error.ts:91](https://github.com/adobe/aio-commerce-sdk/blob/b828858b2e024cee9599e664761b0da5b22f0cd1/packages/aio-commerce-lib-core/source/lib/error/base-error.ts#L91)
+Defined in: base-error.ts:91
 
 Checks if the error is any CommerceSdkErrorBase instance.
