+++ packages/aio-commerce-lib-core/docs/api-reference/error/classes/CommerceSdkValidationError.md
@@ -1,6 +1,6 @@
 # `CommerceSdkValidationError`
 
-Defined in: [lib/error/validation-error.ts:103](https://github.com/adobe/aio-commerce-sdk/blob/b828858b2e024cee9599e664761b0da5b22f0cd1/packages/aio-commerce-lib-core/source/lib/error/validation-error.ts#L103)
+Defined in: validation-error.ts:103
 
 Represents a validation error in the Commerce SDK.
 This error should be thrown when an input does not conform to the expected schema.
@@ -29,7 +29,7 @@ console.log(error.display());
 new CommerceSdkValidationError(message: string, options: CommerceSdkValidationErrorOptions): CommerceSdkValidationError;
 ```
 
-Defined in: [lib/error/validation-error.ts:112](https://github.com/adobe/aio-commerce-sdk/blob/b828858b2e024cee9599e664761b0da5b22f0cd1/packages/aio-commerce-lib-core/source/lib/error/validation-error.ts#L112)
+Defined in: validation-error.ts:112
 
 Constructs a new CommerceSdkValidationError instance.
 
@@ -56,7 +56,7 @@ Constructs a new CommerceSdkValidationError instance.
 readonly issues: GenericIssue[];
 ```
 
-Defined in: [lib/error/validation-error.ts:104](https://github.com/adobe/aio-commerce-sdk/blob/b828858b2e024cee9599e664761b0da5b22f0cd1/packages/aio-commerce-lib-core/source/lib/error/validation-error.ts#L104)
+Defined in: validation-error.ts:104
 
 ---
 
@@ -66,7 +66,7 @@ Defined in: [lib/error/validation-error.ts:104](https://github.com/adobe/aio-com
 readonly optional traceId: string;
 ```
 
-Defined in: [lib/error/base-error.ts:54](https://github.com/adobe/aio-commerce-sdk/blob/b828858b2e024cee9599e664761b0da5b22f0cd1/packages/aio-commerce-lib-core/source/lib/error/base-error.ts#L54)
+Defined in: base-error.ts:54
 
 #### Inherited from
 
@@ -82,7 +82,7 @@ Defined in: [lib/error/base-error.ts:54](https://github.com/adobe/aio-commerce-s
 get fullStack(): string;
 ```
 
-Defined in: [lib/error/base-error.ts:96](https://github.com/adobe/aio-commerce-sdk/blob/b828858b2e024cee9599e664761b0da5b22f0cd1/packages/aio-commerce-lib-core/source/lib/error/base-error.ts#L96)
+Defined in: base-error.ts:96
 
 Returns the full stack trace of the error and its causes.
 
@@ -104,7 +104,7 @@ Returns the full stack trace of the error and its causes.
 get rootCause(): unknown;
 ```
 
-Defined in: [lib/error/base-error.ts:109](https://github.com/adobe/aio-commerce-sdk/blob/b828858b2e024cee9599e664761b0da5b22f0cd1/packages/aio-commerce-lib-core/source/lib/error/base-error.ts#L109)
+Defined in: base-error.ts:109
 
 Returns the root cause of the error.
 
@@ -124,7 +124,7 @@ Returns the root cause of the error.
 display(withColor: boolean): string;
 ```
 
-Defined in: [lib/error/validation-error.ts:124](https://github.com/adobe/aio-commerce-sdk/blob/b828858b2e024cee9599e664761b0da5b22f0cd1/packages/aio-commerce-lib-core/source/lib/error/validation-error.ts#L124)
+Defined in: validation-error.ts:124
 
 Returns a pretty string (and colored) representation of the validation error.
 
@@ -152,7 +152,7 @@ toJSON(): {
 };
 ```
 
-Defined in: [lib/error/base-error.ts:124](https://github.com/adobe/aio-commerce-sdk/blob/b828858b2e024cee9599e664761b0da5b22f0cd1/packages/aio-commerce-lib-core/source/lib/error/base-error.ts#L124)
+Defined in: base-error.ts:124
 
 Converts the error to a JSON-like representation.
 
@@ -210,7 +210,7 @@ traceId: undefined | string;
 toString(inspect: boolean): string;
 ```
 
-Defined in: [lib/error/base-error.ts:139](https://github.com/adobe/aio-commerce-sdk/blob/b828858b2e024cee9599e664761b0da5b22f0cd1/packages/aio-commerce-lib-core/source/lib/error/base-error.ts#L139)
+Defined in: base-error.ts:139
 
 Returns a pretty string representation of the error.
 
@@ -236,7 +236,7 @@ Returns a pretty string representation of the error.
 static isSdkError(error: unknown): error is CommerceSdkErrorBase;
 ```
 
-Defined in: [lib/error/base-error.ts:91](https://github.com/adobe/aio-commerce-sdk/blob/b828858b2e024cee9599e664761b0da5b22f0cd1/packages/aio-commerce-lib-core/source/lib/error/base-error.ts#L91)
+Defined in: base-error.ts:91
 
 Checks if the error is any CommerceSdkErrorBase instance.
