+++ packages/aio-commerce-lib-core/source/index.ts
@@ -1,21 +1,15 @@
-/*
- * Copyright 2025 Adobe. All rights reserved.
- * This file is licensed to you under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License. You may obtain a copy
- * of the License at http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under
- * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
- * OF ANY KIND, either express or implied. See the License for the specific language
- * governing permissions and limitations under the License.
- */
-
 /**
- * This module re-exports all the sub-modules of the package as namespaces.
  * @privateRemarks This annotations are used by Typedoc to correctly generate the API reference. Don't remove them.
  *
  * @module index
  * @mergeModuleWith <project>
  */
 
-export * as Error from "./lib/error";
+/**
+ * @privateRemarks
+ * Dummy file for Typedoc to correctly generate API reference file structure.
+ * **Remove this file** when the package has more than one sub-module.
+ *
+ * (At the time of writing, there's only the `error` sub-module.)
+ */
+
