+++ packages/aio-commerce-lib-core/test/spec/error/base-error.test.ts
@@ -12,9 +12,9 @@
 
 import { describe, expect, it } from "vitest";
 
-import { CommerceSdkErrorBase } from "~/lib/error/base-error";
+import { CommerceSdkErrorBase } from "~/error/base-error";
 
-import type { CommerceSdkErrorBaseOptions } from "~/lib/error/base-error";
+import type { CommerceSdkErrorBaseOptions } from "~/error/base-error";
 
 // Create a concrete implementation for testing
 class TestError extends CommerceSdkErrorBase {
