+++ packages/aio-commerce-lib-core/tsdown.config.ts
@@ -3,5 +3,5 @@ import { defineConfig } from "tsdown";
 
 export default defineConfig({
   ...baseConfig,
-  entry: ["./source/index.ts", "./source/lib/error/index.ts"],
+  entry: ["./source/**/index.ts"],
 });
