+++ packages/aio-commerce-sdk/package.json
@@ -33,16 +33,6 @@
   "module": "./dist/es/index.js",
   "types": "./dist/cjs/index.d.cts",
   "exports": {
-    ".": {
-      "import": {
-        "types": "./dist/es/index.d.ts",
-        "default": "./dist/es/index.js"
-      },
-      "require": {
-        "types": "./dist/cjs/index.d.cts",
-        "default": "./dist/cjs/index.cjs"
-      }
-    },
     "./*": {
       "import": {
         "types": "./dist/es/*.d.ts",
