--- packages/aio-commerce-sdk/source/index.ts
@@ -1,21 +1,0 @@
-/*
- * Copyright 2025 Adobe. All rights reserved.
- * This file is licensed to you under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License. You may obtain a copy
- * of the License at http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under
- * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
- * OF ANY KIND, either express or implied. See the License for the specific language
- * governing permissions and limitations under the License.
- */
-
-/**
- * As much as you might be tempted to change the export path to use the relative files in this package,
- * don't do it. If symbols get re-exported more than once, build tools will not be able to tree-shake them.
- *
- * Use always `export * as <name> from <packageName>`, where packageName is the NPM name, not a relative path.
- */
-
-export * as Auth from "@adobe/aio-commerce-lib-auth";
-export * as Core from "@adobe/aio-commerce-lib-core";
